+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# Nagios Check for HP Procurve Switches
+# Nagios Check for HP ProCurve Switches
+
+## v1.1
+- Check has now been tested on a HP ProCurve 2530-24G and should work just fine for all HP ProCurve Switches
+  (See DEVICES.md for tested support)
+- Slightly changed some checks inner workings. A full refactor is coming soon for easier enhancements and testing.
 
 ## v1.0
-Initial version with tested support for HP Procurve 2510G-24 switches
+- Initial version with tested support for HP ProCurve 2510G-24 switches